+++ Magic/src/main/java/com/elmakers/mine/bukkit/action/builtin/ApplyCooldownAction.java
@@ -120,7 +120,7 @@ public class ApplyCooldownAction extends BaseSpellAction
 	@Override
 	public boolean isUndoable()
 	{
-		return true;
+		return false;
 	}
 
     @Override