+++ Magic/src/main/java/com/elmakers/mine/bukkit/magic/MageClass.java
@@ -32,7 +32,11 @@ public class MageClass extends CasterProperties implements com.elmakers.mine.buk
 
     @Override
     public boolean hasProperty(String key) {
-        return hasOwnProperty(key) || template.hasProperty(key) || mageProperties.hasProperty(key) || (parent != null && parent.hasProperty(key));
+        BaseMagicProperties storage = getStorage(key);
+        if (storage != null) {
+            return storage.hasOwnProperty(key);
+        }
+        return hasOwnProperty(key) || template.hasProperty(key);
     }
 
     @Override
@@ -48,13 +52,17 @@ public class MageClass extends CasterProperties implements com.elmakers.mine.buk
         if (value == null) {
             value = template.getProperty(key);
         }
-        if (value == null && parent != null) {
-            value = parent.getProperty(key);
+        return value;
+    }
+
+    @Override
+    public Object getInheritedProperty(String key) {
+        Object value = super.getProperty(key);
+        if (value == null) {
+            value = template.getProperty(key);
         }
-        // Don't need to check the mage if we have a parent, since the parent
-        // checks for us.
-        if (value == null && parent == null) {
-            value = mageProperties.getProperty(key);
+        if (value == null && parent != null) {
+            value = parent.getInheritedProperty(key);
         }
         return value;
     }