+++ Magic/src/main/java/com/elmakers/mine/bukkit/utility/ConfigurationUtils.java
@@ -9,7 +9,7 @@ import com.elmakers.mine.bukkit.api.magic.MageController;
 import com.elmakers.mine.bukkit.api.spell.PrerequisiteSpell;
 import com.elmakers.mine.bukkit.api.spell.SpellKey;
 import com.elmakers.mine.bukkit.effect.SoundEffect;
-import com.elmakers.mine.bukkit.magic.ParameterizedConfiguration;
+import com.elmakers.mine.bukkit.magic.AttributableConfiguration;
 import de.slikey.effectlib.util.ConfigUtils;
 import de.slikey.effectlib.util.ParticleEffect;
 
@@ -274,8 +274,8 @@ public class ConfigurationUtils extends ConfigUtils {
 
     public static ConfigurationSection cloneConfiguration(ConfigurationSection section)
     {
-        ConfigurationSection copy = section instanceof ParameterizedConfiguration ?
-                new ParameterizedConfiguration((ParameterizedConfiguration)section) :
+        ConfigurationSection copy = section instanceof AttributableConfiguration ?
+                new AttributableConfiguration((AttributableConfiguration)section) :
                 new MemoryConfiguration();
         return addConfigurations(copy, section);
     }